Add files via upload
Assignment 2 
Names: Bwalya Boldwin
Course: DIT 400
S/ID: 2300391
Apology:
Due to network problem  and not having much knowledge this site/application in terms of how to upload the file and inviting the collaborators has lead me to be a late /a bad student who seems not to be following the instruction. Kindly may you find a place in your heart for forgiveness and next time this won't repeat again . Thank! may Jehova god bless and be kind to you.